Move `move_dir()` to its own directory.
